requirements.txt
# Core
pillow
requests
numpy

# WebSocket-Client für lokalen Stream
websocket-client

# Dev-Modus
flask

# Tests
pytest
